+++ service/src/main/java/com/kap/core/dto/mp/mpb/MPBCompanyDTO.java
@@ -1,7 +1,6 @@
 package com.kap.core.dto.mp.mpb;
 
 import com.kap.core.dto.BaseDTO;
-import com.kap.core.dto.wb.wbb.WBBACompanyDTO;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.*;
 
@@ -152,7 +151,7 @@ public class MPBCompanyDTO extends BaseDTO {
     private String nm;
 
     @Schema(title = "점수")
-    private Integer score;
+    private Double score;
 
     @Schema(title = "SQ 평가년도")
     private Integer year;
